pyproject.toml
[project]
name = "pluppkalender"
version = "0.1.0"
description = "An object recoginition model to convert Naturskolan's physical calendar into digital data"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "click>=8.3.1",
    "keras==2.13.*",
    "pillow>=12.0.0",
    "scipy>=1.15.3",
    "tensorflow==2.13.*",
]

[project.scripts]
train_model = "pluppkalender.pluppkalender:train_model"
split_into_days = "pluppkalender.pluppkalender:split_into_days"
predict_days = "pluppkalender.pluppkalender:predict_days"

[tool.uv]
package = true
